refactor: 메서드 분리 (2단계 코드 리뷰 반영)
- winningCount()
- profitRate()
